# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee Tracking API",
        "description": "Dokumentasi API untuk demo tracking lokasi karyawan (Laravel + Sanctum).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server utama"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user baru",
                "operationId": "ca7c0cb6772de9bffbb8c896980cebc8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jimmy"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jimmy@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 6,
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Register success"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jimmy"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jimmy@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcDEF...token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user dan dapatkan token",
                "operationId": "c373bc4cffa4a56fc64e3979b8ee7b96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jimmy@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login success"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jimmy"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jimmy@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcDEF...token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/locations": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Ambil semua lokasi terakhir user",
                "operationId": "2d235517e13dbf0d125785d4a9e5e94f",
                "parameters": [
                    {
                        "name": "max_accuracy",
                        "in": "query",
                        "description": "Filter maksimum akurasi (meter), contoh: 50 → hanya tampil akurasi ≤ 50 m",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar lokasi user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_name": {
                                                        "type": "string",
                                                        "example": "Jimmy"
                                                    },
                                                    "lat": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": -0.947812
                                                    },
                                                    "lng": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 100.417523
                                                    },
                                                    "recorded_at": {
                                                        "type": "string",
                                                        "example": "2025-10-23 20:00:00"
                                                    },
                                                    "accuracy_m": {
                                                        "type": "integer",
                                                        "example": 12,
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/locations/{id}": {
            "put": {
                "tags": [
                    "Location"
                ],
                "summary": "Update atau create lokasi terbaru user (latest only)",
                "operationId": "77b62b3520b16ebb3c3bb7ad55a579a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dummy untuk memenuhi resource route (tidak digunakan di server)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lat",
                                    "lng"
                                ],
                                "properties": {
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -0.947812
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.417523
                                    },
                                    "recorded_at": {
                                        "type": "string",
                                        "example": "2025-10-23T20:00:00+07:00"
                                    },
                                    "accuracy_m": {
                                        "type": "integer",
                                        "example": 10,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lokasi berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "location": {
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_name": {
                                                    "type": "string",
                                                    "example": "Jimmy"
                                                },
                                                "lat": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": -0.947812
                                                },
                                                "lng": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100.417523
                                                },
                                                "recorded_at": {
                                                    "type": "string",
                                                    "example": "2025-10-23 20:00:00"
                                                },
                                                "accuracy_m": {
                                                    "type": "integer",
                                                    "example": 10,
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Location",
            "description": "Location"
        }
    ]
}